Refactor convergence settings to separate sub-advanced-settings module
