+++ src/aiidalab_qe/app/configuration/advanced/convergence/__init__.py
@@ -1,0 +1,7 @@
+from .convergence import ConvergenceConfigurationSettingsPanel
+from .model import ConvergenceConfigurationSettingsModel
+
+__all__ = [
+    "ConvergenceConfigurationSettingsModel",
+    "ConvergenceConfigurationSettingsPanel",
+]
